+++ NEWS.md
@@ -1,3 +1,8 @@
+# roxyglobals 0.2.2
+
+- @autoglobal ignore assignment expressions (#7)
+- @autoglobal search package imports for defined globals (#8)
+
 # roxyglobals 0.2.1
 
 - @autoglobal matching names in parent environments